reorganize nftvm unit tests
