+++ examples/nftvm/chaintest/action_test_helpers.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2024, Ava Labs, Inc. All rights reserved.
 // See the file LICENSE for licensing terms.
 
-package unit_test
+package chaintest
 
 import (
 	"context"
